requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
celery==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
Pillow==11.1.0
pypdfium2==4.30.0
python-decouple==3.8
gunicorn==23.0.0
django-celery-results==2.5.1
numpy>=1.26.0
scikit-image>=0.24.0
opencv-python-headless>=4.10.0
scipy>=1.13.0
